pyproject.toml
[tool.poetry]
name = "fairest"
version = "0.1.1"
description = "Proofreading tool for legal and other documents"
license = "GPL-3.0-or-later"
readme = "README.rst"
authors = ["Ang Houfu <damon.houf@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
pluggy = "^0.13.1"
python-docx = "^0.8.10"
spacy-nightly = "^3.0.0-rc.2"
textstat = "^0.7.0"
en-core-web-sm = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0a0/en_core_web_sm-3.0.0a0.tar.gz"}
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^6"
pytest-cov = "^2.10.1"
pytest-mock = "^3.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
